tantest/requirements.txt
# Project dependencies
claude-agent-sdk==0.1.3
python-dotenv==1.1.1
